+++ react/EXPERIMENTAL_useTableBulkActions.ts
@@ -1,3 +1,3 @@
 export {
   default,
-} from './components/EXPERIMENTAL_Table/stateContainers/bulkActions'
+} from './components/EXPERIMENTAL_Table/hooks/useTableBulkActions'
