+++ react/EXPERIMENTAL_useTableVisibility.ts
@@ -1,3 +1,3 @@
 export {
   default,
-} from './components/EXPERIMENTAL_Table/stateContainers/visibility'
+} from './components/EXPERIMENTAL_Table/hooks/useTableVisibility'
