+++ react/components/EXPERIMENTAL_Table/README.md
@@ -56,7 +56,7 @@ Different than the previous version the `Table v2` is completely stateless, mean
 
 ```js
 // Imports
-const useTableMeasures = require('./stateContainers/measures.tsx').default
+const useTableMeasures = require('./hooks/useTableMeasures.tsx').default
 const Tag = require('../Tag/index.js').default
 
 // Define the columns
@@ -149,7 +149,7 @@ function StateHookExample() {
 # Clickable rows
 
 ```js
-const useTableMeasures = require('./stateContainers/measures.tsx').default
+const useTableMeasures = require('./hooks/useTableMeasures.tsx').default
 
 const columns = [
   {
@@ -207,7 +207,7 @@ function ClickExample() {
 
 ```js
 // Imports
-const useTableMeasures = require('./stateContainers/measures.tsx').default
+const useTableMeasures = require('./hooks/useTableMeasures.tsx').default
 const Toggle = require('../Toggle/index.js').default
 
 const columns = [
@@ -267,7 +267,7 @@ function LoadingExample() {
 
 ```js
 // Imports
-const useTableMeasures = require('./stateContainers/measures.tsx').default
+const useTableMeasures = require('./hooks/useTableMeasures.tsx').default
 const Toggle = require('../Toggle/index.js').default
 const Spinner = require('../Spinner/index.js').default
 
@@ -331,8 +331,8 @@ function CustomLoadingExample() {
 
 ```js
 // Imports
-const useTableMeasures = require('./stateContainers/measures.tsx').default
-const useTableVisibility = require('./stateContainers/visibility.ts').default
+const useTableMeasures = require('./hooks/useTableMeasures.tsx').default
+const useTableVisibility = require('./hooks/useTableVisibility.ts').default
 
 const columns = [
   {
@@ -371,8 +371,8 @@ Empty states can also be customized, the passed children will be rendered inside
 
 ```js
 // Imports
-const useTableMeasures = require('./stateContainers/measures.tsx').default
-const useTableVisibility = require('./stateContainers/visibility.ts').default
+const useTableMeasures = require('./hooks/useTableMeasures.tsx').default
+const useTableVisibility = require('./hooks/useTableVisibility.ts').default
 const Button = require('../Button/index.js').default
 
 const columns = [
@@ -422,7 +422,7 @@ function CustomEmptyStateExample() {
 
 ```js
 // Imports
-const useTableMeasures = require('./stateContainers/measures.tsx').default
+const useTableMeasures = require('./hooks/useTableMeasures.tsx').default
 
 // Define the columns
 const columns = [
@@ -606,8 +606,8 @@ Check the console when selecting/unselecting rows or clicking an action button i
 
 ```js
 // Imports
-const useTableMeasures = require('./stateContainers/measures.tsx').default
-const useTableBulkActions = require('./stateContainers/bulkActions.tsx').default
+const useTableMeasures = require('./hooks/useTableMeasures.tsx').default
+const useTableBulkActions = require('./hooks/useTableBulkActions.tsx').default
 
 // Define the columns
 const columns = [
@@ -713,8 +713,8 @@ This feature creates a last extra column with an ActionMenu component per line.
 
 ```js
 // Imports
-const useTableLineActions = require('./stateContainers/lineActions.tsx').default
-const useTableMeasures = require('./stateContainers/measures.tsx').default
+const useTableLineActions = require('./hooks/useTableLineActions.tsx').default
+const useTableMeasures = require('./hooks/useTableMeasures.tsx').default
 
 // Define the columns
 const columns = [
@@ -799,7 +799,7 @@ function LineActionsExample() {
 # Filter Bar
 
 ```js
-const useTableMeasures = require('./stateContainers/measures.tsx').default
+const useTableMeasures = require('./hooks/useTableMeasures.tsx').default
 const Input = require('../Input').default
 const Checkbox = require('../Checkbox').default
 
@@ -1139,8 +1139,8 @@ function FilterBarExample() {
 
 ```js
 // Imports
-const useTableMeasures = require('./stateContainers/measures.tsx').default
-const useTableVisibility = require('./stateContainers/visibility.ts').default
+const useTableMeasures = require('./hooks/useTableMeasures.tsx').default
+const useTableVisibility = require('./hooks/useTableVisibility.ts').default
 
 // Define the columns
 const columns = [
@@ -1319,7 +1319,7 @@ The `UNSAFE_InputCustom` provides a simple way of passing a custom input to the 
 
 ```js
 // Imports
-const useTableMeasures = require('./stateContainers/measures.tsx').default
+const useTableMeasures = require('./hooks/useTableMeasures.tsx').default
 const Input = require('../Input/index.js').default
 
 /** Define the columns */
