+++ react/components/EXPERIMENTAL_Table/hooks/useTableVisibility.ts
@@ -1,5 +1,5 @@
 import { useMemo, useState, useCallback } from 'react'
-import { Column, Items } from '../index'
+import { Column } from '../index'
 
 export default function useTableVisibility({ columns }: VisibilityData) {
   const [hiddenColumns, setHiddenColumns] = useState(getHiddenColumns(columns))
